+++ bindings/vue/src/components/VOnsIcon.vue
@@ -1,0 +1,42 @@
+<template>
+  <ons-icon>
+  </ons-icon>
+</template>
+
+<script>
+  export default {
+    //--------------------------------
+    // lifecycle hooks
+    //--------------------------------
+
+    // beforeCreate() {
+    // },
+
+    // created() {
+    // },
+
+    // beforeMount() {
+    // },
+
+    // mounted() {
+    // },
+
+    // beforeUpdate() {
+    // },
+
+    // updated() {
+    // },
+
+    // activated() {
+    // },
+
+    // deactivated() {
+    // },
+
+    // beforeDestroy() {
+    // },
+
+    // destroyed() {
+    // },
+  };
+</script>