+++ bindings/vue/src/components/index.js
@@ -1,3 +1,4 @@
 export { default as VOnsButton } from './VOnsButton.vue';
+export { default as VOnsIcon } from './VOnsIcon.vue';
 export { default as VOnsRange } from './VOnsRange.vue';
 export { default as VOnsSwitch } from './VOnsSwitch.vue';